refactor(tui): extract token and startup runtime component
